Separate Audio and VADAudio classes into separate packages
